sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ordo-sdk"
version = "0.3.0"
description = "Python SDK for Ordo Rule Engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Ordo Team" }]
keywords = ["rule-engine", "ordo", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "requests>=2.28",
]

[project.optional-dependencies]
grpc = [
    "grpcio>=1.50",
    "protobuf>=4.21",
]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "responses>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["ordo"]
